+++ src/main/java/refinedstorage/gui/GuiProcessingPatternEncoder.java
@@ -5,6 +5,7 @@ import net.minecraft.init.SoundEvents;
 import refinedstorage.RefinedStorage;
 import refinedstorage.container.ContainerProcessingPatternEncoder;
 import refinedstorage.network.MessageGridPatternCreate;
+import refinedstorage.network.MessageProcessingPatternEncoderClear;
 import refinedstorage.tile.TileProcessingPatternEncoder;
 
 import java.io.IOException;
@@ -26,8 +27,12 @@ public class GuiProcessingPatternEncoder extends GuiBase {
     public void update(int x, int y) {
     }
 
-    public boolean isHoveringOverCreatePattern(int mouseX, int mouseY) {
+    private boolean isOverCreatePattern(int mouseX, int mouseY) {
         return inBounds(152, 38, 16, 16, mouseX, mouseY) && processingPatternEncoder.canCreatePattern();
+    }
+
+    private boolean isOverClear(int mouseX, int mouseY) {
+        return inBounds(136, 75, 7, 7, mouseX, mouseY);
     }
 
     @Override
@@ -38,7 +43,7 @@ public class GuiProcessingPatternEncoder extends GuiBase {
 
         int ty = 0;
 
-        if (isHoveringOverCreatePattern(mouseX - guiLeft, mouseY - guiTop)) {
+        if (isOverCreatePattern(mouseX - guiLeft, mouseY - guiTop)) {
             ty = 1;
         }
 
@@ -54,8 +59,12 @@ public class GuiProcessingPatternEncoder extends GuiBase {
         drawString(7, 7, t("gui.refinedstorage:processing_pattern_encoder"));
         drawString(7, 78, t("container.inventory"));
 
-        if (isHoveringOverCreatePattern(mouseX, mouseY)) {
+        if (isOverCreatePattern(mouseX, mouseY)) {
             drawTooltip(mouseX, mouseY, t("gui.refinedstorage:processing_pattern_encoder.pattern_create"));
+        }
+
+        if (isOverClear(mouseX, mouseY)) {
+            drawTooltip(mouseX, mouseY, t("misc.refinedstorage:clear"));
         }
     }
 
@@ -63,8 +72,12 @@ public class GuiProcessingPatternEncoder extends GuiBase {
     protected void mouseClicked(int mouseX, int mouseY, int mouseButton) throws IOException {
         super.mouseClicked(mouseX, mouseY, mouseButton);
 
-        if (isHoveringOverCreatePattern(mouseX - guiLeft, mouseY - guiTop)) {
+        if (isOverCreatePattern(mouseX - guiLeft, mouseY - guiTop)) {
             RefinedStorage.INSTANCE.network.sendToServer(new MessageGridPatternCreate(processingPatternEncoder.getPos().getX(), processingPatternEncoder.getPos().getY(), processingPatternEncoder.getPos().getZ()));
+
+            mc.getSoundHandler().playSound(PositionedSoundRecord.getMasterRecord(SoundEvents.UI_BUTTON_CLICK, 1.0F));
+        } else if (isOverClear(mouseX - guiLeft, mouseY - guiTop)) {
+            RefinedStorage.INSTANCE.network.sendToServer(new MessageProcessingPatternEncoderClear(processingPatternEncoder));
 
             mc.getSoundHandler().playSound(PositionedSoundRecord.getMasterRecord(SoundEvents.UI_BUTTON_CLICK, 1.0F));
         }